pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "gitbrief"
version = "0.1.0"
description = "Your daily developer standup - powered by your Git history"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "Akshat190", email = "akshatchoksi8@gmail.com"}
]
keywords = ["cli", "git", "developer-tools", "ai", "productivity", "ollama"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Version Control :: Git",
    "Topic :: Utilities",
]
requires-python = ">=3.10"
dependencies = [
    "typer>=0.12.0",
    "rich>=13.0.0",
    "GitPython>=3.1.0",
    "requests>=2.31.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[project.scripts]
gitbrief = "cli:app"

[project.urls]
Homepage = "https://github.com/Akshat190/gitbrief"
Repository = "https://github.com/Akshat190/gitbrief"
Issues = "https://github.com/Akshat190/gitbrief/issues"

[tool.setuptools]
packages = ["cli", "git_reader", "summarizer", "prompts", "utils"]

[tool.setuptools.package-data]
gitbrief = ["*.json"]

[tool.black]
line-length = 100
target-version = ["py310"]

[tool.ruff]
line-length = 100
target-version = "py310"
